Move `ScanCode` to `keyCodes.ts`
