+++ src/vs/base/common/keybindingParser.ts
@@ -3,10 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { KeyCodeUtils } from 'vs/base/common/keyCodes';
-import { ChordKeybinding, Keybinding, SimpleKeybinding } from 'vs/base/common/keybindings';
+import { KeyCodeUtils, ScanCodeUtils } from 'vs/base/common/keyCodes';
+import { ChordKeybinding, Keybinding, SimpleKeybinding, ScanCodeBinding } from 'vs/base/common/keybindings';
 import { OperatingSystem } from 'vs/base/common/platform';
-import { ScanCodeBinding, ScanCodeUtils } from 'vs/base/common/scanCode';
 
 export class KeybindingParser {
