+++ src/vs/platform/keyboardLayout/common/keyboardLayout.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Event } from 'vs/base/common/event';
-import { ScanCode, ScanCodeUtils } from 'vs/base/common/scanCode';
+import { ScanCode, ScanCodeUtils } from 'vs/base/common/keyCodes';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IKeyboardEvent } from 'vs/platform/keybinding/common/keybinding';
 import { DispatchConfig } from 'vs/platform/keyboardLayout/common/dispatchConfig';
